+++ nn/conv.go
@@ -9,13 +9,13 @@ import (
 )
 
 type conv2d struct {
-	InChannels  int
-	OutChannels int
-	KernelSize  int
-	Stride      int
-	Padding     int
-	Dilation    int
-	Groups      int
+	InChannels  int64
+	OutChannels int64
+	KernelSize  int64
+	Stride      int64
+	Padding     int64
+	Dilation    int64
+	Groups      int64
 	PaddingMode string
 	Weight      torch.Tensor
 	Bias        torch.Tensor
@@ -25,7 +25,7 @@ type conv2d struct {
 // TODO(qijun): only support zero padding mode
 // only support symmetry kernel/stride/padding/dilation
 func Conv2d(inChannels, outChannels, kernelSize, stride, padding, dilation,
-	groups int, bias bool, paddingMode string) Module {
+	groups int64, bias bool, paddingMode string) Module {
 	c := &conv2d{
 		InChannels:  inChannels,
 		OutChannels: outChannels,
@@ -36,11 +36,11 @@ func Conv2d(inChannels, outChannels, kernelSize, stride, padding, dilation,
 		Groups:      groups,
 		PaddingMode: "zeros",
 	}
-	c.Weight = torch.Empty([]int{outChannels, inChannels / groups, kernelSize,
+	c.Weight = torch.Empty([]int64{outChannels, inChannels / groups, kernelSize,
 		kernelSize}, true)
 	initializer.KaimingUniform(&c.Weight, math.Sqrt(5.0), "fan_in", "leaky_relu")
 	if bias {
-		c.Bias = torch.Empty([]int{outChannels}, true)
+		c.Bias = torch.Empty([]int64{outChannels}, true)
 		fanIn, _ := initializer.CalculateFanInAndFanOut(c.Weight)
 		bound := 1.0 / math.Sqrt(float64(fanIn))
 		initializer.Uniform(&c.Bias, -bound, bound)
@@ -50,6 +50,6 @@ func Conv2d(inChannels, outChannels, kernelSize, stride, padding, dilation,
 
 // Forward method
 func (c *conv2d) Forward(x torch.Tensor) torch.Tensor {
-	return functional.Conv2d(x, c.Weight, c.Bias, []int{c.Stride, c.Stride},
-		[]int{c.Padding, c.Padding}, []int{c.Dilation, c.Dilation}, c.Groups)
+	return functional.Conv2d(x, c.Weight, c.Bias, []int64{c.Stride, c.Stride},
+		[]int64{c.Padding, c.Padding}, []int64{c.Dilation, c.Dilation}, c.Groups)
 }
